+++ src/components/ProofOfAddress/Guidance/graphic.js
