+++ src/components/ProofOfAddress/Guidance/index.tsx
