+++ src/components/ProofOfAddress/Guidance/style.scss
