+++ src/components/ProofOfAddress/PoAIntro/index.tsx
