+++ src/components/ProofOfAddress/PoAIntro/style.scss
